requirements.txt
openai-whisper==20231117
twilio==8.12.0
sounddevice==0.4.6
scipy==1.12.0
numpy==1.26.3
tabulate==0.9.0
streamlit==1.32.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0 